Feature: Adicionando operações para o controller OperacaoController.java

+++ joao-figueredo-dr1-tp1/src/main/java/infnet/edu/br/joaofigueredodr1tp1/dto/OperacaoDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class OperacaoDTO {
-    private Long id;
-    private double num1;
-    private double num2;
+    private Double id;
+    private Double num1;
+    private Double num2;
 }
