+++ joao-figueredo-dr1-tp1/src/main/java/infnet/edu/br/joaofigueredodr1tp1/repository/IOperacaoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IOperacaoRepository  extends JpaRepository<Operacao, Long> {
+public interface IOperacaoRepository extends JpaRepository<Operacao, Double> {
 }
